crud jabatan & seed factory sebagian

+++ database/migrations/2022_12_04_133232_create_finances_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string("description");
             $table->integer("budget");
-            $table->decimal("percentage");
+            $table->double("percentage");
             $table->timestamps();
         });
     }
